f #479: 1. add status column to the project table 2. center all tables pagination in all tabs 3. remove banned and draft from listing column in projects table 4. remove unnecessary styles and codes and decompose big components into smaller ones 5. replace user website with user email address in my account head 6. remove copy user wallet address icon 7. disable edit and view buttons in projects table for cancelled projects 8. improve my account responsiveness, specifically tables 9. remove networksObj and replace with networksParams
