+++ src/components/ExternalLink.tsx
@@ -1,16 +1,27 @@
 import React, { ReactElement } from 'react';
+import styled from 'styled-components';
 
 const ExternalLink = (props: {
 	children?: ReactElement[] | ReactElement;
 	href: string;
-	title?: string;
+	title?: string | JSX.Element;
+	color?: string;
 }) => {
-	const { children, href, title } = props;
+	const { children, href, title, color } = props;
 	return (
-		<a href={href} rel='noopener noreferrer' target='_blank'>
+		<Styled
+			color={color}
+			href={href}
+			rel='noopener noreferrer'
+			target='_blank'
+		>
 			{title || children}
-		</a>
+		</Styled>
 	);
 };
+
+const Styled = styled.a`
+	color: ${props => props.color || 'inherit'};
+`;
 
 export default ExternalLink;
