+++ src/components/views/userPublicProfile/PublicProfileContributes.tsx
@@ -4,13 +4,12 @@ import styled from 'styled-components';
 import { brandColors, neutralColors, P } from '@giveth/ui-design-system';
 
 import { Flex } from '@/components/styled-components/Flex';
-import PublicProfileDonationsTab from './PublicProfileDonationsTab';
+import PublicProfileDonationsTab from './donationsTab/PublicProfileDonationsTab';
 import PublicProfileLikedTab from './PublicProfileLikedTab';
-import PublicProfileProjectsTab from './PublicProfileProjectsTab';
+import PublicProfileProjectsTab from './projectsTab/PublicProfileProjectsTab';
 import PublicProfileOverviewTab from './PublicProfileOverviewTab';
 import { IUserPublicProfileView } from './UserPublicProfile.view';
 import { Container } from '@/components/Grid';
-import { mediaQueries } from '@/lib/constants/constants';
 
 enum EPublicProfile {
 	OVERVIEW,
@@ -54,109 +53,79 @@ const PublicProfileContributes: FC<IUserPublicProfileView> = ({
 	}, [router?.query?.tab]);
 
 	return (
-		<PubliCProfileTabsAndProjectContainer>
-			<ProfileContainer>
-				<PubliCProfileTabsContainer>
-					{myAccount && (
-						<PubliCProfileTab
-							active={tab === EPublicProfile.OVERVIEW}
-							onClick={() => setTab(EPublicProfile.OVERVIEW)}
-						>
-							Overview
-						</PubliCProfileTab>
+		<ProfileContainer>
+			<PublicProfileTabsContainer>
+				{myAccount && (
+					<PublicProfileTab
+						active={tab === EPublicProfile.OVERVIEW}
+						onClick={() => setTab(EPublicProfile.OVERVIEW)}
+					>
+						Overview
+					</PublicProfileTab>
+				)}
+				<PublicProfileTab
+					active={tab === EPublicProfile.PROJECTS}
+					onClick={() => setTab(EPublicProfile.PROJECTS)}
+				>
+					{`${myAccount ? 'My ' : user.name + '’s'} projects`}
+					{myAccount && user?.projectsCount != 0 && (
+						<Count active={tab === EPublicProfile.PROJECTS}>
+							{user?.projectsCount}
+						</Count>
 					)}
-					<PubliCProfileTab
-						active={tab === EPublicProfile.PROJECTS}
-						onClick={() => setTab(EPublicProfile.PROJECTS)}
-					>
-						{`${myAccount ? 'My ' : user.name + '’s'} projects`}
-						{myAccount && user?.projectsCount != 0 && (
-							<Count active={tab === EPublicProfile.PROJECTS}>
-								{user?.projectsCount}
-							</Count>
-						)}
-					</PubliCProfileTab>
-					<PubliCProfileTab
-						active={tab === EPublicProfile.DONATIONS}
-						onClick={() => setTab(EPublicProfile.DONATIONS)}
-					>
-						{`${myAccount ? 'My ' : ''}Donations`}
-						{myAccount && user?.donationsCount != 0 && (
-							<Count active={tab === EPublicProfile.DONATIONS}>
-								{user?.donationsCount}
-							</Count>
-						)}
-					</PubliCProfileTab>
-					<PubliCProfileTab
-						active={tab === EPublicProfile.LIKED}
-						onClick={() => setTab(EPublicProfile.LIKED)}
-					>
-						Liked projects
-						{myAccount && !!user.likedProjectsCount && (
-							<Count active={tab === EPublicProfile.LIKED}>
-								{user?.likedProjectsCount}
-							</Count>
-						)}
-					</PubliCProfileTab>
-				</PubliCProfileTabsContainer>
-				{tab === EPublicProfile.OVERVIEW && (
-					<PublicProfileOverviewTab
-						user={user}
-						myAccount={myAccount}
-					/>
-				)}
-				{tab === EPublicProfile.PROJECTS && (
-					<PublicProfileProjectsTab
-						user={user}
-						myAccount={myAccount}
-					/>
-				)}
-				{tab === EPublicProfile.DONATIONS && (
-					<DonationsTableWrapper>
-						<PublicProfileDonationsTab
-							user={user}
-							myAccount={myAccount}
-						/>
-					</DonationsTableWrapper>
-				)}
-				{tab === EPublicProfile.LIKED && (
-					<PublicProfileLikedTab user={user} myAccount={myAccount} />
-				)}
-			</ProfileContainer>
-		</PubliCProfileTabsAndProjectContainer>
+				</PublicProfileTab>
+				<PublicProfileTab
+					active={tab === EPublicProfile.DONATIONS}
+					onClick={() => setTab(EPublicProfile.DONATIONS)}
+				>
+					{`${myAccount ? 'My ' : ''}Donations`}
+					{myAccount && user?.donationsCount != 0 && (
+						<Count active={tab === EPublicProfile.DONATIONS}>
+							{user?.donationsCount}
+						</Count>
+					)}
+				</PublicProfileTab>
+				<PublicProfileTab
+					active={tab === EPublicProfile.LIKED}
+					onClick={() => setTab(EPublicProfile.LIKED)}
+				>
+					Liked projects
+					{myAccount && !!user.likedProjectsCount && (
+						<Count active={tab === EPublicProfile.LIKED}>
+							{user?.likedProjectsCount}
+						</Count>
+					)}
+				</PublicProfileTab>
+			</PublicProfileTabsContainer>
+			{tab === EPublicProfile.OVERVIEW && (
+				<PublicProfileOverviewTab user={user} myAccount={myAccount} />
+			)}
+			{tab === EPublicProfile.PROJECTS && (
+				<PublicProfileProjectsTab user={user} myAccount={myAccount} />
+			)}
+			{tab === EPublicProfile.DONATIONS && (
+				<PublicProfileDonationsTab user={user} myAccount={myAccount} />
+			)}
+			{tab === EPublicProfile.LIKED && (
+				<PublicProfileLikedTab user={user} myAccount={myAccount} />
+			)}
+		</ProfileContainer>
 	);
 };
 
-export default PublicProfileContributes;
-
-const PubliCProfileTabsAndProjectContainer = styled.div``;
-
 const ProfileContainer = styled(Container)`
-	${mediaQueries.mobileL} {
-		padding: 0 10px 0 10px !important;
-	}
-	${mediaQueries.mobileM} {
-		padding: 0 10px 0 10px !important;
-	}
-	${mediaQueries.mobileS} {
-		padding: 0 10px 0 10px !important;
-	}
+	padding: 0 !important;
 `;
 
-const PubliCProfileTabsContainer = styled(Flex)`
+const PublicProfileTabsContainer = styled(Flex)`
 	padding: 37px 0;
 	gap: 16px;
 	max-width: 600px;
 	overflow: auto;
 `;
 
-const DonationsTableWrapper = styled.div`
-	overflow: auto;
-`;
-
-const PubliCProfileTab = styled(P)<ITab>`
+const PublicProfileTab = styled(P)<ITab>`
 	display: flex;
-	flex-direction: row;
 	align-items: center;
 	padding: 9px 10px;
 	word-break: break-word;
@@ -185,3 +154,5 @@ const Count = styled.div`
 	font-size: 12px;
 	margin-left: 4px;
 `;
+
+export default PublicProfileContributes;
